requirements.txt
haystack-ai==2.2.4
genson==1.3.0
neo4j==5.22.0
pymongo==4.8.0
python-dotenv==1.0.1
streamlit==1.36.0
